+++ ydb/core/persqueue/common/key.h
