+++ ydb/core/persqueue/pqtablet/blob/blob.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <ydb/core/persqueue/key.h>
+#include <ydb/core/persqueue/common/key.h>
 
 #include <util/datetime/base.h>
 #include <util/generic/size_literals.h>
