+++ ydb/core/persqueue/pqtablet/partition/partition.h
@@ -5,7 +5,7 @@
 
 #include <ydb/core/persqueue/pqtablet/blob/blob.h>
 #include <ydb/core/persqueue/pqtablet/blob/header.h>
-#include <ydb/core/persqueue/key.h>
+#include <ydb/core/persqueue/common/key.h>
 #include "partition_init.h"
 #include "partition_sourcemanager.h"
 #include "partition_types.h"
