+++ ydb/core/persqueue/pqtablet/partition/partition_blob_encoder.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <ydb/core/persqueue/pqtablet/blob/blob.h>
-#include <ydb/core/persqueue/key.h>
+#include <ydb/core/persqueue/common/key.h>
 #include <ydb/core/persqueue/events/internal.h>
 
 namespace NKikimr::NPQ {
