+++ ydb/core/persqueue/pqtablet/partition/partition_compactification.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ydb/core/persqueue/key.h>
+#include <ydb/core/persqueue/common/key.h>
 #include <ydb/core/persqueue/pqtablet/blob/blob.h>
 
 #include <ydb/core/keyvalue/keyvalue_events.h>
