+++ ydb/core/persqueue/pqtablet/partition/sourceid.h
@@ -3,7 +3,7 @@
 #include <ydb/core/keyvalue/keyvalue_events.h>
 #include <ydb/core/persqueue/heartbeat.h>
 #include <ydb/core/persqueue/sourceid_info.h>
-#include <ydb/core/persqueue/key.h>
+#include <ydb/core/persqueue/common/key.h>
 #include "ownerinfo.h"
 #include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
 #include <ydb/core/protos/pqconfig.pb.h>
