+++ ydb/core/persqueue/ut/sourceid_ut.cpp
@@ -1,7 +1,7 @@
 #include <ydb/core/persqueue/pqtablet/partition/sourceid.h>
 
 #include <ydb/core/keyvalue/keyvalue_events.h>
-#include <ydb/core/persqueue/key.h>
+#include <ydb/core/persqueue/common/key.h>
 
 #include <library/cpp/testing/unittest/registar.h>
