+++ ydb/public/sdk/cpp/src/client/topic/ut/topic_to_table_ut.cpp
@@ -9,7 +9,7 @@
 
 #include <ydb/core/cms/console/console.h>
 #include <ydb/core/keyvalue/keyvalue_events.h>
-#include <ydb/core/persqueue/key.h>
+#include <ydb/core/persqueue/common/key.h>
 #include <ydb/core/persqueue/pqtablet/blob/blob.h>
 #include <ydb/core/persqueue/events/global.h>
 #include <ydb/core/persqueue/pqtablet/cache/pq_l2_service.h>
